Move `try_recv` onto `Chan`
This follows the same idea as on `Sender`.
